revert bits besides bridge to .ts files
