+++ core/src/util.ts
@@ -1,3 +1,5 @@
+import { WindowCapacitor } from "./definitions-internal";
+
 export enum ExceptionCode {
   /**
    * API is not implemented.
@@ -22,3 +24,34 @@ export class CapacitorException extends Error {
     super(message);
   }
 }
+
+export const getPlatformId = (win: WindowCapacitor) => {
+  if (win.androidBridge) {
+    return 'android';
+  } else if (
+    win.webkit &&
+    win.webkit.messageHandlers &&
+    win.webkit.messageHandlers.bridge
+  ) {
+    return 'ios';
+  } else {
+    return 'web';
+  }
+};
+
+export const convertFileSrcServerUrl = (webviewServerUrl: string, filePath: string) => {
+  if (typeof filePath === 'string') {
+    if (filePath.startsWith('/')) {
+      return webviewServerUrl + '/_capacitor_file_' + filePath;
+    } else if (filePath.startsWith('file://')) {
+      return (
+        webviewServerUrl + filePath.replace('file://', '/_capacitor_file_')
+      );
+    } else if (filePath.startsWith('content://')) {
+      return (
+        webviewServerUrl + filePath.replace('content:/', '/_capacitor_content_')
+      );
+    }
+  }
+  return filePath;
+};
